refactor: enhanced structure of tests and test data
